Reorganize modbus plugin code to factor out time-conversions and config handling.
